+++ lib/contentful/management/entry.rb
@@ -71,8 +71,8 @@ module Contentful
       # @param [String] wanted_locale
       #
       # @return [Hash] localized fields
-      def fields(wanted_locale = default_locale)
-        requested_locale = locale || wanted_locale
+      def fields(wanted_locale = locale)
+        requested_locale = wanted_locale || default_locale
         @fields[requested_locale] = {} unless @fields[requested_locale]
 
         default_fields = @fields[default_locale] || {}